useTransitionSetting
