+++ packages/hooks/src/setting/index.ts
@@ -2,3 +2,4 @@ export * from './useRootSetting'
 export * from './useMenuSetting'
 export * from './useHeaderSetting'
 export * from './useMultipleTabSetting'
+export * from './useTransitionSetting'
